requirements.txt
mysql-connector-python
bluepy
